+++ cpp/color.cpp
@@ -1,0 +1,17 @@
+//
+// Created by Shohei Matsugu on 2017/07/10.
+//
+
+#include "../h/color.h"
+
+color::color(){
+    this->r = 0;
+    this->g = 0;
+    this->b = 0;
+}
+
+color::color(int r, int g, int b){
+    this->r = r;
+    this->g = g;
+    this->b = b;
+}
